third_party/blink/web_tests/fast/forms/parser-associated-form-removal-expected.txt

Form controls associated with a form by the parser should be reset when that form leaves their tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Removing a form from the document while leaving its associated element in...
PASS document.querySelector("input").form is null


Removing a form and its associated element...
PASS table2.querySelector("input").form is non-null.
...and then removing the form.
PASS table2.querySelector("input").form is null
PASS successfullyParsed is true

TEST COMPLETE
